requirements.txt
Flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
redis==5.0.1
Flask-Session==0.5.0
python-dotenv==1.0.0
requests==2.31.0 